requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.2
django-cors-headers==4.0.0
django-rest-authtoken==2.1.4
django-sslserver==0.22
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.8.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
